Relocate css values and rename interpolators configs dir to props dir
