+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/interpolation/styles/AnimationStyleInterpolator.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <reanimated/CSS/config/interpolators/registry.h>
+#include <reanimated/CSS/config/props/registry.h>
 #include <reanimated/CSS/interpolation/groups/RecordPropertiesInterpolator.h>
 
 #include <memory>
