+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/interpolation/transforms/TransformOperationInterpolator.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include <reanimated/CSS/common/values/CSSValue.h>
 #include <reanimated/CSS/interpolation/transforms/TransformInterpolator.h>
 #include <reanimated/CSS/interpolation/transforms/TransformOperation.h>
+#include <reanimated/CSS/values/CSSValue.h>
 
 #include <memory>
 #include <string>
