+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/util/props.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <reanimated/CSS/common/definitions.h>
-#include <reanimated/CSS/config/interpolators/registry.h>
+#include <reanimated/CSS/config/props/registry.h>
 
 #include <memory>
 #include <string>
