+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/values/base/CSSColor.cpp
@@ -1,4 +1,4 @@
-#include <reanimated/CSS/common/values/CSSColor.h>
+#include <reanimated/CSS/values/base/CSSColor.h>
 
 namespace reanimated::css {
