+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/values/base/CSSColor.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <reanimated/CSS/common/definitions.h>
-#include <reanimated/CSS/common/values/CSSValue.h>
+#include <reanimated/CSS/values/CSSValue.h>
 
 #include <worklets/Tools/JSISerializer.h>
