+++ packages/react-native-reanimated/src/css/platform/native/style/processors/svg/others.ts
@@ -1,7 +1,7 @@
 'use strict';
-
 import type { ValueProcessor } from '../../types';
 
-export const processStripUnit: ValueProcessor<number | string, number> = (
-  value
-) => (typeof value === 'number' ? value : parseFloat(value));
+export const convertPercentageToNumber: ValueProcessor<
+  number | string,
+  number
+> = (value) => (typeof value === 'number' ? value : parseFloat(value) / 100);
